project/requirements.txt
flask
flask-login
flask-migrate
flask-script
flask-wtf
flask-sqlalchemy
werkzeug
click==7.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
email_validator
PyJWT